+++ src/main/java/be/baur/sds/ComplexType.java
@@ -1,18 +1,17 @@
 package be.baur.sds;
 
 import be.baur.sda.Node;
-import be.baur.sda.NodeSet;
 import be.baur.sds.common.Attribute;
 import be.baur.sds.common.Component;
-import be.baur.sds.model.AbstractGroup;
 
 /**
  * A <code>ComplexType</code> represents an SDS definition of a complex SDA
  * node. It is a container for other components and/or model groups.
  */
-public class ComplexType extends ComponentType {
+//@Deprecated
+public class ComplexType extends NodeType {
 
-	private ComplexType globalcomplextype = null; // the global complex type this component refers to.
+	//private ComplexType globalcomplextype = null; // the global complex type this component refers to.
 	
 	/** Creates a complex type with the specified <code>name</code>.*/
 	public ComplexType(String name) {
@@ -29,30 +28,26 @@ public class ComplexType extends ComponentType {
 	 * unexpected behavior at some point in the future, but we shall cross that
 	 * bridge when we get there.
 	 */
-	@Override
-	public NodeSet getNodes() {
-		
-		if (getGlobalType() == null) return super.getNodes();
-		
-		if (globalcomplextype == null) // not bound yet, so get it from the schema root
-			globalcomplextype = (ComplexType) this.root().getNodes().get(getGlobalType()).get(1);
-		
-		return globalcomplextype != null ? globalcomplextype.getNodes() : new NodeSet();
-	}
+//	@Override
+//	public NodeSet getNodes() {
+//		
+//		if (getGlobalType() == null) return super.getNodes();
+//		
+//		if (globalcomplextype == null) // not bound yet, so get it from the schema root
+//			globalcomplextype = (ComplexType) this.root().getNodes().get(getGlobalType()).get(1);
+//		
+//		return globalcomplextype != null ? globalcomplextype.getNodes() : new NodeSet();
+//	}
 	
 	
-	public final Node toNode() {
+	public Node toNode() {
 		
-		Node node; // resulting node, returned at the end of this method
+		Node node = new Node(Component.NODE.tag);
 		
-		// Omit the name for model groups or type references
-		// with the same name as the referenced global type.
-		if (! (this instanceof AbstractGroup) ) {
-			node = new Node(Component.NODE.tag);
-			if (getGlobalType() == null || ! getName().equals(getGlobalType()))
-				node.setValue(getName());
-		}
-		else node = new Node(getName());
+		// Omit the name for type references with the same name as the referenced global type.
+
+		if (getGlobalType() == null || ! getName().equals(getGlobalType()))
+			node.setValue(getName());
 		
 		if (getGlobalType() != null) // Render the type attribute if we have one.
 			node.add(new Node(Attribute.TYPE.tag, getGlobalType()));
@@ -68,8 +63,8 @@ public class ComplexType extends ComponentType {
 	}
 	
 	
-	@Override
-	public final String toString() {
-		return toNode().toString();
-	}
+//	@Override
+//	public final String toString() {
+//		return toNode().toString();
+//	}
 }