+++ src/main/java/be/baur/sds/model/SequenceGroup.java
@@ -15,10 +15,10 @@ import be.baur.sds.common.Component;
  * </pre>
  * Which means that a last name is expected, optionally preceded by a middle name.
  */
-public final class Group extends AbstractGroup {
+public final class SequenceGroup extends ModelGroup {
 
 	/** Creates a group. */
-	public Group() {
+	public SequenceGroup() {
 		super(Component.GROUP.tag);
 	}
 }